refactor(#1414): extract zstd frame parser to zstd_frame module; keep chunk_decompressor under campsite threshold
Move the authoritative zstd frame-header Dictionary_ID parser and the
metadata+frame dictionary-rejection helper into a new
storage/sstable/zstd_frame.rs module (with its own unit test), so
chunk_decompressor.rs stays under the 800-line campsite threshold. fmt applied.

verify.rs (pre-existing 2542-line campsite violation) still grows by the
required VerifyErrorClass::UnsupportedCompressionFeature variant + classification
+ tests; splitting it is out of scope for this P0 fix (epic #1116) — acknowledged
via CQLITE_ALLOW_FILE_GROWTH.
